refactor: 자동으로 생긴 spec파일들 test 폴더로 위치이동
